Created all migrations

+++ database/migrations/2021_04_24_165830_create_friends_table.php
@@ -15,8 +15,8 @@ class CreateFriendsTable extends Migration
     {
         Schema::create('friends', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unsigned()->nullable();
-            $table->integer('friend_id')->unsigned()->nullable();
+            $table->bigInteger('user_id')->unsigned()->nullable();
+            $table->bigInteger('friend_id')->unsigned()->nullable();
             $table->foreign('user_id')->references('id')->on('users')->onUpdate('set null')->onDelete('set null');
             $table->foreign('friend_id')->references('id')->on('users')->onUpdate('set null')->onDelete('set null');
             $table->timestamps();
